Move data normalization to api service
Refs #46.
