+++ app/services/api.js
@@ -1,5 +1,6 @@
+import capitalize from 'lodash/capitalize';
+import keyBy from 'lodash/keyBy';
 import queryString from 'query-string';
-
 
 export function get(url, queryParams) {
   const query = queryString.stringify(queryParams);
@@ -8,8 +9,55 @@ export function get(url, queryParams) {
   return window.fetch(fullUrl).then(response => response.json());
 }
 
+function normalizeOwner(owner) {
+  switch (owner.toLowerCase()) {
+    case 'avain asumisoikeus':
+    case 'avain asumisoikeus oy':
+      return 'AVAIN Asumisoikeus Oy';
+    case 'asuntosäätiön asumisoikeus oy':
+      return 'Asuntosäätiön Asumisoikeus Oy';
+    case 'helsingin seudun asumisoikeusyhdistys':
+    case 'helsingin seudun asumisoikeusyhdistys helas':
+      return 'Helsingin Seudun Asumisoikeusyhdistys HELAS';
+    case 'ta-asumisoikeus':
+    case 'ta-asumisoikeus oy':
+      return 'TA-Asumisoikeus Oy';
+    default:
+      return owner;
+  }
+}
+
+function getUnitOwnerAndName(data) {
+  const dataName = data.name.fi;
+  const slashIndex = dataName.indexOf('/');
+  if (slashIndex !== -1) {
+    const owner = normalizeOwner(dataName.substring(0, slashIndex));
+    const name = dataName.substring(slashIndex + 1);
+    return { owner, name };
+  }
+  return {
+    owner: normalizeOwner(dataName),
+    name: data.street_address.fi,
+  };
+}
+
+function getUnitData(data) {
+  const { owner, name } = getUnitOwnerAndName(data);
+  const [longitude, latitude] = data.location.coordinates;
+  return {
+    id: data.id,
+    name,
+    owner,
+    streetAddress: data.street_address.fi,
+    addressZip: data.address_zip,
+    city: capitalize(data.municipality),
+    url: data.www_url.fi,
+    coordinates: { longitude, latitude },
+  };
+}
+
 function getUnits() {
-  return get(
+  const request = get(
     'unit/',
     {
       service: '25304',
@@ -17,7 +65,8 @@ function getUnits() {
       page: '1',
       page_size: '1000',
     }
-  ).then(data => data.results);
+  );
+  return request.then(data => data.results).then(units => keyBy(units.map(getUnitData), 'id'));
 }
 
 export default { getUnits };
